Add spdx license identifiers
